\name{CDF}
\alias{CDF}
\title{Cumulative distribution function}
\description{Evaluates the cumulative distribution function using the integral of its density function.}
\usage{CDF(c, density)}
\arguments{
  \item{c}{Value at which the CDF is to be evaluated.}
  \item{density}{Density function.}
}
\author{
Stanislas Hubeaux, \email{stanislas.hubeaux@alumni.epfl.ch} 

Kaspar Rufibach, \email{kaspar.rufibach@gmail.com} \cr \url{http://www.kasparrufibach.ch}
}

\note{Function not intended to be invoked by the user.}

\keyword{methods}
